agamotto/requirements.txt
google-ads
google-cloud
google-cloud-bigquery
google-cloud-storage
grpcio
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
pandas
Flask
dask
tensorflow
tensorflow-datasets
numpy
matplotlib

